put StepManager into separate file
