+++ backend/protzilla/run.py
@@ -14,7 +14,8 @@ from backend.protzilla.constants.data_types import Connection
 import backend.protzilla.constants.paths as paths
 from backend.protzilla.constants.date_format import metadata_date_format
 from backend.protzilla.form import Form
-from backend.protzilla.steps import Messages, Output, Plots, Step, StepManager, Section
+from backend.protzilla.steps import Messages, Output, Plots, Step, Section
+from backend.protzilla.step_manager import StepManager
 from backend.protzilla.utilities import format_trace
